# Loops everywhere - Cleanup init in GetFrame and Degrain SSE2 code